fix: Ensure message fields are copied when building retry request
Use the to_dict() function when building the ReadRowsRequest arguments.

This way, all available fields in the message will get copied over.
The `message.filter` field needed a special handling, since the to_dict() function doesn't seem to parse the value the way we want it to be.

Fixes internal bug #214449800